Create AdventAPI module to fetch input from the server
Created folders Common, input and src:
* Common for now contains AdventAPI.hs, code to manage reading the
puzzle inputs
* input contains all the puzzle inputs
* src now contains all the source files

AdventAPI's readInput tries to read the input from a file in the input
folder. If the requested file does not exist, it will fetch the input
from Advent of Code's servers and store it in a file. This should make
sure each input is only fetched from the server once. For this to work,
there must be a file containing your session cookie, as the inputs vary
per account.