Setup tests using doctests
So this took a long time to figure out. The most important detail is
that "write-ghc-environment-files: always" bit in cabal.project, as
without it, tests would always fail due to not finding the advent
module.

Also moved aoc2020.cabal to src/2020, so that when other years start
being worked on, all that is required is to add a new cabal file. This
may still require a bit of work, as executables with the same name might
cause issues. Perhaps not, given that they can be disambiguated by doing
"cabal run aoc2020:day01", for example.

Furthermore, forced strictness in day15.hs and day25.hs, to avoid a
couple of stack overflows that occurred when running those executables
in doctest.